+++ Changes.md
@@ -1,10 +1,20 @@
 # Changes Logs 📜
 
-## 1. Changed to BookPasal 📖
+## 1\. Changed to BookPasal 📖
 
 ```
 changed to bookpasal
 ```
 
-- Changed BookCycle to BookPasal **(name conflict)**. 🥲
+- Changed ~BookCycle~ to BookPasal **(name conflict)**. 🥲
 - Modifed Readme from template 📃
+
+## 2\. Added cart system 🛒
+
+```
+added cart
+```
+
+- added cart 🛒
+- card state management and context
+- bug fixes on card link